Add basic docker-compose file
In the docker-compose, we run the plants-manager service with a mongodb
container.
The DB credentials are passed as environment variables to both the
service and to mongoDB.
We created a volume to be able to store our data persistently
For the service to be able to communicate with the DB, we set a shared
network for both containers.