Store plants in mongodb instead of in-memory
I chose to use MongoDB since I would like to have the ability to easily
add more features and data types to the database.
Since we won't work with complex queries and we are not concerned about
data consistency (as long as its eventually consistent) MongoDB seemed to be suitable.

As all plant names are unique, and probably most queries will include
searching a plant's details (i.e more read than write operations),
I chose to use a unique index on the name property.

In the model module we initiate a client for our MongoDB instance
In init_db() we initiate the plants DB: A plants collection and the name index.
Notice that the function is idempotent